+++ open_spiel/games/catch_test.cc
@@ -60,7 +60,7 @@ void GetAllStatesTest() {
 
   // Verify normalized observation matches string represtation.
   for (const auto& pair : states) {
-    std::vector<double> obs;
+    std::vector<float> obs;
     pair.second->ObservationTensor(0, &obs);
     const std::string& str = pair.first;
     SPIEL_CHECK_EQ(obs.size(), str.size() - kDefaultRows);
