+++ open_spiel/games/clobber.h
@@ -87,7 +87,7 @@ class ClobberState : public State {
   std::string InformationStateString(Player player) const override;
   std::string ObservationString(Player player) const override;
   void ObservationTensor(Player player,
-                         std::vector<double>* values) const override;
+                         std::vector<float>* values) const override;
   std::unique_ptr<State> Clone() const override {
     return std::unique_ptr<State>(new ClobberState(*this));
   }
