+++ open_spiel/games/coop_box_pushing.cc
@@ -516,7 +516,7 @@ int CoopBoxPushingState::ObservationPlane(std::pair<int, int> coord,
 }
 
 void CoopBoxPushingState::ObservationTensor(Player player,
-                                            std::vector<double>* values) const {
+                                            std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
   if (fully_observable_) {
