+++ open_spiel/games/lewis_signaling.cc
@@ -113,7 +113,7 @@ std::string LewisSignalingState::ObservationString(Player player) const {
 }
 
 void LewisSignalingState::ObservationTensor(Player player,
-                                            std::vector<double>* values) const {
+                                            std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
