+++ open_spiel/games/negotiation.cc
@@ -176,7 +176,7 @@ std::vector<int> NegotiationGame::ObservationTensorShape() const {
 }
 
 void NegotiationState::ObservationTensor(Player player,
-                                         std::vector<double>* values) const {
+                                         std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
