+++ open_spiel/games/phantom_ttt.cc
@@ -152,8 +152,8 @@ std::string PhantomTTTState::InformationStateString(Player player) const {
       + ActionSequenceToString(player);
 }
 
-void PhantomTTTState::InformationStateTensor(
-    Player player, std::vector<double>* values) const {
+void PhantomTTTState::InformationStateTensor(Player player,
+                                             std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
 
@@ -203,7 +203,7 @@ std::string PhantomTTTState::ObservationString(Player player) const {
 }
 
 void PhantomTTTState::ObservationTensor(Player player,
-                                        std::vector<double>* values) const {
+                                        std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
   values->resize(game_->ObservationTensorSize());
