+++ open_spiel/games/universal_poker.h
@@ -70,9 +70,9 @@ class UniversalPokerState : public State {
   std::string InformationStateString(Player player) const override;
   std::string ObservationString(Player player) const override;
   void InformationStateTensor(Player player,
-                              std::vector<double> *values) const override;
+                              std::vector<float> *values) const override;
   void ObservationTensor(Player player,
-                         std::vector<double> *values) const override;
+                         std::vector<float> *values) const override;
   std::unique_ptr<State> Clone() const override;
 
   // The probability of taking each possible action in a particular info state.
